==> chapters/04-analysis.Rmd <==
# Auswertung {#total-crashes}

> Zur besseren Lesbarkeit wird der R Code in diesem Kapitel nicht gezeigt. Dieser besteht größtenteils aus Plots und ist bis auf wenige Ausnahmen nicht weiter relevant. 

```{r echo = FALSE}
knitr::opts_chunk$set(
  echo = FALSE
)
```

```{r}
library(tidyverse)
library(patchwork)
library(sf)
library(viridis)
```

```{r}
load(
  here::here("data/regensburg_data.rda")
)

load(
  here::here("data/shapefiles.rda")
)
```

Im Stadtgebiet Regensburg geschahen von den Jahren `r min(data$year)` bis `r max(data$year)` insgesamt `r nrow(data)` Unfälle mit Personenschaden. Abbildung \@ref(fig:unfaelle-timeseries) zeigt die monatlichen Unfälle in diesem Zeitraum.

```{r unfaelle-timeseries, fig.cap = "Monatliche Unfälle in Regensburg."}
data |> 
  count(date) |> 
  ggplot(aes(x = date, y = n)) +
  geom_line() +
  scale_x_date(
    date_breaks = "6 months",
    date_labels = "%m/%Y",
    limits = c(min(data$date), max(data$date))
  ) +
  labs(
    title = "Monatliche Verkehrsunfälle mit Personenschaden",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    x = "Datum",
    y = "Anzahl an Unfällen"
  )
```

Während sich kein eindeutiger Auf- oder Abwärtstrend feststellen lässt, zeigen die Daten dennoch eine Jährliche Periodizität: Im Sommer finden die meisten Unfälle mit Personenschaden statt, während die Anzahl der Unfälle von Herbst bis Frühjahr sinkt.

```{r}
RelativeChange <- function(start_year, end_year) {
  
  a <- nrow(filter(data, year == end_year)) 
  b <- nrow(filter(data, year == start_year))
 
  res <- (1 - a/b) * 100
  
  return(res)
}
```

Abbildung \@ref(fig:unfaelle-jaehrlich) zeigt die Anzahl der jährlichen Unfälle in Regensburg. Im Jahr 2020 zeigt sich ein Rückgang von `r round(RelativeChange(2019, 2020), digits = 0)`%. Dieser könnte auf geringeren Verkehr aufgrund der Corona Pandemie zurückgeführt werden. Dies wird durch Abbildung \@ref(fig:unfaelle-jaehrlich-typ) verdeutlicht: Alle dokumentierten Verkehrsmittel hatten einen Rückgang der jährlichen Unfälle von 2019 bis 2020.

```{r unfaelle-jaehrlich, cache = TRUE, fig.cap = "Jährliche Unfälle mit Personenschaden."}
data |> 
  count(year, severity) |> 
  ggplot() +
  geom_col(
    aes(x = year, y = n, fill = severity)
  ) +
  labs(
    title = "Jährliche Unfälle mit Personenschaden",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    x = "Jahr",
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  scale_fill_custom
```

Abbildung \@ref(fig:unfaelle-jaehrlich-typ) zeigt zudem, dass sich die Anzahl der Unfälle aller Verkehrsmittel außer Fahrrad auf einem fallenden Trend befinden. Die Anzahl der Unfälle mit Fahrradbeteiligung dagegen stieg bis 2020 kontinuierlich an.

```{r unfaelle-jaehrlich-typ, cache = TRUE, fig.cap = "Jährliche Unfälle mit Personenschaden nach Verkehrsteilnehmer unterteilt."}
data |> 
  pivot_longer(
    cols = bicycle:bike,
    names_to = "type_of_vehicle",
    values_to = "took_part"
  ) |> 
  mutate(
    type_of_vehicle = as_factor(type_of_vehicle) |> 
      fct_recode(
        "Fahrrad" = "bicycle",
        "Auto" = "car",
        "Fußgänger" = "pedestrian",
        "Motorrad" = "bike"
      )
  ) |> 
  filter(
    took_part == TRUE
  ) |> 
  count(year, severity, type_of_vehicle) |> 
  
  ggplot() +
  geom_col(aes(x = year, y = n, fill = severity)) +
  labs(
    title = "Jährliche Unfälle nach Teilnehmer",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    x = "Jahr",
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  scale_fill_custom +
  facet_wrap(
    facets = vars(type_of_vehicle),
    scales = "free_y"
  )
```

Der wichtigste Faktor, der die Unfälle unterscheidet, ist die Schwere der schlimmsten Verletzung (Leicht, Schwer, Tödlich). Ungefähr `r round(nrow(filter(data, severity == "Leicht")) / nrow(data), digits = 2) * 100`% Unfälle gehen mit höchstens leichten Verletzungen aus. Lediglich `r round(nrow(filter(data, severity == "Toedlich")) / nrow(data), digits = 2) * 100`% der Unfälle enden tödlich.

```{r}
data |> 
  filter(severity == "Toedlich") |> 
  select(-id, -truck, -lng, -lat, -datetime, -date) |> 
  rename(
    "Jahr" = "year",
    "Monat" = "month", 
    "Zeit" = "hour",
    "Tag" = "weekday",
    "Schwere" = "severity",
    "Licht" = "light_condition",
    "Straße" = "street_condition",
    "Fahrrad" = "bicycle",
    "Auto" = "car",
    "Fußgänger" = "pedestrian",
    "Motorrad" = "bike"
  ) |> 
  DT::datatable(
    caption = "Unfälle mit Todefolge."
  )
```

## Unfälle nach Monat

```{r}
data.barplot <- 
  data |> 
  mutate(
    month = as_factor(month) |> 
      fct_recode(
        "Jan" = "1",
        "Feb" = "2",
        "Mär" = "3",
        "Apr" = "4",
        "Mai" = "5",
        "Jun" = "6",
        "Jul" = "7",
        "Aug" = "8",
        "Sep" = "9",
        "Okt" = "10",
        "Nov" = "11",
        "Dez" = "12"
      ),
    severity = fct_infreq(severity) |> 
      fct_rev(),
    light_condition = fct_infreq(light_condition) |> 
      fct_rev(),
    street_condition = fct_infreq(street_condition) |> 
      fct_rev()
    )
```

Abbildung \@ref(fig:unfaelle-monthly) zeigt die monatliche Verteilung der dokumentierten Unfälle. Es lässt sich ablesen, dass die Anzahl der Unfälle vom Winter bis Sommer ansteigt und in Juli ein Maximum annimmt. Mit dem Beginn der Sommerferien in Bayern im August sinkt die Anzahlt der Unfälle dann um `r round(1 - nrow(filter(data.barplot, month == "Aug")) / nrow(filter(data.barplot, month == "Jul")), digits = 2) * 100`%.

```{r, fig.height = 6}
plot.bar.month_1 <- 
  data.barplot |> 
  count(month, severity) |> 
  ggplot(aes(x = month, y = n)) +
  geom_col(aes(fill = severity)) +
  labs(
    x = "Monat",
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )

plot.bar.month_2 <- 
  data.barplot |> 
  count(month, light_condition) |> 
  ggplot(aes(x = month, y = n)) +
  geom_col(
    aes(fill = light_condition)
  ) +
  labs(
    x = "Monat",
    y = "Anzahl an Unfällen",
    fill = "Lichtverhältnisse"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )

plot.bar.month_3 <- 
  data.barplot |> 
  count(month, street_condition) |> 
  ggplot(aes(x = month, y = n)) +
  geom_col(
    aes(fill = street_condition)
  ) +
  labs(
    x = "Datum",
    y = "Anzahl an Unfällen",
    fill = "Straßenverhältnisse"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )
```

```{r unfaelle-monthly, fig.height = 6, fig.cap = "Monatiliche Unfälle mit Personenschaden. A) Schwere: Über das Jahr hinweg gibt es kaum relative Unterschiede. B) Lichtverhältnisse: Über die Wintermonate nehmen die Unfälle bei Dunkelheit zu. C) Die Unfälle bei Nässe nehmen von Herbst bis Winter zu. Vom Frühjahr bis Sommer nimmt die Anzahl an Unfällen bei Nässe ab."}
plot.bar.month_1 / 
  plot.bar.month_2 / 
  plot.bar.month_3 + 
  plot_annotation(
    title = "Monatliche Unfälle mit Personenschaden",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    tag_levels = "A",
    tag_suffix = "  "
  ) &
  scale_fill_custom &
  theme(
    axis.text.x = element_text(
      angle = 45,
      vjust = 0.6
    )
  )
```

Die Schwere der Unfälle unterscheidet sich zwischen den Jahreszeiten kaum.

## Unfälle nach Uhrzeit

```{r, fig.height = 6}
plot.bar.time_1 <- 
  data.barplot |> 
  count(hour, severity) |> 
  ggplot(aes(x = hour, y = n)) +
  geom_col(
    aes(fill = severity)
  ) +
  labs(
    x = "Datum",
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )

plot.bar.time_2 <- 
  data.barplot |> 
  count(hour, light_condition) |> 
  ggplot(aes(x = hour, y = n)) +
  geom_col(
    aes(fill = light_condition)
  ) +
  labs(
    x = "Datum",
    y = "Anzahl an Unfällen",
    fill = "Lichtverhältnisse"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )

plot.bar.time_3 <- 
  data.barplot |> 
  count(hour, street_condition) |> 
  ggplot(aes(x = hour, y = n)) +
  geom_col(
    aes(fill = street_condition)
  ) +
  labs(
    x = "Datum",
    y = "Anzahl an Unfällen",
    fill = "Straßenverhältnisse"
  ) +
  theme(
    plot.background = element_rect(
      colour = "white"
    )
  )
```

Die meisten Unfälle mit Personenschaden finden tagsüber statt. Zusätzlich zeigt Abbildung \@ref(fig:unfaelle-time) zwei Peaks: Um sieben Uhr und von 16-17 Uhr. Diese stimmen mit dem Arbeitsverkehr überein.

Abbildung \@ref(fig:unfaelle-time) zeigt auch, dass Unfälle bei Dunkelheit zwischen 16 und 6 Uhr stattfinden. Das ist keine überraschende Erkenntnis, stärkt aber das Vertrauen in die Zuverlässigkeit der Daten.

```{r unfaelle-time, fig.height = 6, fig.cap = "Unfälle nach Tageszeit."}
plot.bar.time_1 / 
  plot.bar.time_2 / 
  plot.bar.time_3 +
  plot_annotation(
    title = "Unfälle mit Personenschaden nach Tagesezeit",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    tag_levels = "A",
    tag_suffix = "  "
  ) &
  scale_fill_viridis(
    discrete = TRUE,
    direction = -1,
    option = "D"
  )
```

## Wo finden die meisten Unfälle statt?

```{r}
sf.data <-
  data |> 
  st_as_sf(coords = c("lng", "lat"), crs = "WGS84") |> 
  rename(
    points = geometry
  ) |> 
  st_join(
    y = sf.districts |> rename("district_shape" = geometry),
    join = st_within,
    left = TRUE
  ) |> 
  select(-m2)
```

```{r}
data.district_plot <-
  sf.data |> 
  as_tibble() |> 
  left_join(
    y = sf.districts |> rename("district_polygon" = geometry) ,
    by = "district"
  ) |>
  drop_na(district) |>
  add_count(district) |> 
  mutate(
    district = as_factor(district) |> 
      fct_infreq() |> 
      fct_rev()
  ) |> 
  select(district, district_polygon, n, m2) |> 
  unique()
```

```{r}
plot.district_total_bar <- 
  data.district_plot |> 
  ggplot(aes(x = district, y = n)) +
  geom_bar(
    aes(fill = n),
    stat = "identity"
  ) +
  labs(
    y = "Anzahl an Unfällen (N)",
    fill = "N"
  ) +
  theme(
    axis.title.y = element_blank(),
    plot.background = element_rect(
      size = 0
    )
  ) +
  coord_flip()
```

```{r}
plot.district_total_map <-
  data.district_plot |>
  ggplot() +
  geom_sf(aes(geometry = district_polygon, fill = n), lwd = 0) + 
  geom_sf(data = sf.districts, alpha = 0) +
  geom_sf(data = sf.regensburg, alpha = 0, lwd = 0.9, linetype = 1) +
  ggthemes::theme_map() +
  theme(
    legend.position = "none"
  )
```

Abbildung \@ref(fig:unfaelle-district1) zeigt die Anzahl an Unfällen von `r min(data$year)` und `r max(data$year)`. In diesem Zeitraum haben sich die meisten Unfälle in der Regensburger Innenstadt (`r filter(data.district_plot, district == "Innenstadt")$n`) zugetragen.

Berechnet man die Unfälle pro Quadratkilometer, wird der Unterschied zwischen der Innenstadt und den Umliegenden Stadtteilen noch eindeutiger (siehe Abbildung \@ref(fig:unfaelle-district2)).

```{r unfaelle-district1, fig.cap = "Unfälle mit Personenschaden pro Stadtteil (Abolute Zahlen)."}
plot.district_total_bar + 
  inset_element(
    plot.district_total_map, 
    left = 0.35, 
    bottom = 0, 
    right = 1, 
    top = 0.65
  ) +
  plot_annotation(
    title = "Unfälle nach Ortsteil",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})")
  ) & 
  scale_fill_viridis(option = "D")
```

```{r}
data.district_m2_plot <- 
  data.district_plot |> 
  unique() |> 
  mutate(
    n_per_m2 = n/m2 * 10^6,
    district = reorder(district, n_per_m2)
  )
```

```{r}
plot.district_m2_bar <- 
  data.district_m2_plot |> 
  ggplot(aes(x = district, y = n_per_m2)) +
  geom_bar(
    aes(fill = n_per_m2),
    stat = "identity"
  ) +
  labs(
    y = "Unfälle pro Quadratkilometer (N/qkm)",
    fill = "N/qkm"
  ) +
  theme(
    axis.title.y = element_blank(),
    plot.background = element_rect(
      size = 0
    )
  ) +
  coord_flip()
```

```{r}
plot.district_m2_map <-
  data.district_m2_plot |>
  ggplot() +
  geom_sf(aes(geometry = district_polygon, fill = n_per_m2), lwd = 0) + 
  geom_sf(data = sf.districts, alpha = 0) +
  geom_sf(data = sf.regensburg, alpha = 0, lwd = 0.9, linetype = 1) +
  ggthemes::theme_map() +
  theme(
    legend.position = "none"
  )
```

```{r unfaelle-district2, fig.cap = "Unfälle mit Personenschaden pro Quadratkilometer."}
plot.district_m2_bar +
  inset_element(
    plot.district_m2_map,
    left = 0.35,
    right = 1,
    bottom = 0,
    top = 0.65
  ) +
  plot_annotation(
    title = "Unfälle pro Quadratkilometer nach Ortsteil",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})")
  ) & 
  scale_fill_viridis(option = "D")
```

## Welches ist das gefährlichste Verkehrsmittel

Betrachtet man die absolute Anzahl der Unfälle mit Personenschaden (siehe Abbildung \@ref(fig:who1)), kommt der Eindruck zustande, dass das Auto das gefährlichste Verkehrsmittel ist. Normiert man die Daten (sieh Abbildung \@ref(fig:who2)), zeigt sich dagegen, dass das Auto relativ sicher ist. So besitzt es das geringste Verhältnis zwischen schweren/tödlichen und leichten Unfällen. 

```{r who1, fig.cap = "Unfälle nach Verkehrsteilnehmer."}
data |> 
  select(severity, bicycle:bike) |> 
  pivot_longer(-severity) |> 
  filter(value == TRUE) |> 
  mutate(
    name = as_factor(name) |> 
      fct_infreq() |> 
      fct_recode(
        "Auto" = "car",
        "Fahrrad" = "bicycle",
        "Motorrad" = "bike",
        "Fußgänger" = "pedestrian"
      )
  ) |> 
  count(severity, name) |> 
  ggplot() + 
  geom_col(aes(x = name, y = n, fill = severity)) +
  labs(
    title = "Unfälle nach Verkehrsteilnehmer",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  scale_fill_custom +
  theme(axis.title.x = element_blank())
```

```{r who2, fig.cap = "Unfälle nach Verkehrsteilnehmer (normiert)."}
data |> 
  select(severity, bicycle:bike) |> 
  pivot_longer(-severity) |> 
  filter(value == TRUE) |> 
  mutate(
    name = as_factor(name) |> 
      fct_infreq() |> 
      fct_recode(
        "Auto" = "car",
        "Fahrrad" = "bicycle",
        "Motorrad" = "bike",
        "Fußgänger" = "pedestrian"
      )
  ) |> 
  count(severity, name) |> 
  ggplot() + 
  geom_col(aes(x = name, y = n, fill = severity), position = position_fill()) +
  labs(
    title = "Unfälle nach Verkehrsteilnehmer",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    y = "Anzahl an Unfällen",
    fill = "Schwere"
  ) +
  scale_fill_custom +
  theme(axis.title.x = element_blank())
```

In Relation zur Gesamtanzahl der Unfälle mit Personenschaden der einzelnen Verkehrsmittel, sind an den meisten tödlichen Unfällen Motorradfahrer beteiligt (siehe \@ref(fig:who3)).

```{r who3, fig.cap = "Anteil der tödlichen Unfälle nach Verkehrsteilnehmer."}
data |> 
  select(severity, bicycle:bike) |> 
  pivot_longer(-severity) |> 
  filter(value == TRUE) |> 
  mutate(
    name = as_factor(name) |> 
      fct_infreq() |> 
      fct_recode(
        "Auto" = "car",
        "Fahrrad" = "bicycle",
        "Motorrad" = "bike",
        "Fußgänger" = "pedestrian"
      )
  ) |> 
  add_count(name) |> 
  rename("n1" = n) |> 
  count(severity, name, n1) |> 
  rename("n2" = n) |>
  mutate(
    n = n2/n1
  ) |>
  filter(severity == "Toedlich") |>
  ggplot() +
  geom_col(aes(x = name, y = n, fill = "severity")) +
  labs(
    title = "Anteil der tödlichen Unfälle",
    subtitle = glue::glue("Regensburg ({min(data$year)} - {max(data$year)})"),
    y = "Anteil an Unfällen"
  ) +
  scale_y_continuous(
    labels = scales::label_percent()
  ) +
  scale_fill_custom +
  theme(
    axis.title.x = element_blank(),
    legend.position = "none"
  )
```

Trotz der geringeren Gesamtanzahl an Unfällen, ist das Motorrad somit das gefährlichste Verkehrsmittel.
